requirements.txt
boto3=1.21.16
python-dotenv=0.19.2